restaurant2:
Add a second back button but only visible when a menu ingredient is pulled up. The button will take user back to menu builder instead of the previous page like the go back button. The new back button can have text on it that says back to menu.

Add a thin crust ingredient checkbox to pizza.